refactor plugins and configurations
