+++ frontend/src/components/Site/ObservedPropertyModal.vue
@@ -65,7 +65,7 @@
 
 <script lang="ts">
 import { ref } from 'vue'
-import axios from '@/axios.config'
+import axios from '@/plugins/axios.config'
 import { useDataStore } from '@/store/data'
 
 export default {
