+++ frontend/src/components/Site/ProcessingLevelModal.vue
@@ -50,7 +50,7 @@
 </template>
 
 <script lang="ts">
-import axios from '@/axios.config'
+import axios from '@/plugins/axios.config'
 import { ref } from 'vue'
 import { useDataStore } from '@/store/data'
