+++ frontend/src/components/Site/SensorModal.vue
@@ -94,7 +94,7 @@
 
 <script lang="ts">
 import { computed, ref } from 'vue'
-import axios from '@/axios.config'
+import axios from '@/plugins/axios.config'
 import { useDataStore } from '@/store/data'
 
 export default {
