+++ frontend/src/main.ts
@@ -1,15 +1,8 @@
 import { createApp } from 'vue'
+import { store } from './store'
 import App from './App.vue'
 import router from './router/router'
-import { store } from './store'
-
-import '@mdi/font/css/materialdesignicons.css'
-import 'vuetify/styles'
-import { createVuetify } from 'vuetify'
-import * as components from 'vuetify/components'
-import * as directives from 'vuetify/directives'
-
-const vuetify = createVuetify({ components, directives })
+import vuetify from '@/plugins/vuetify'
 
 const app = createApp(App)
 app.use(router)
