+++ frontend/src/plugins/vuetify.ts
@@ -1,0 +1,33 @@
+import '@mdi/font/css/materialdesignicons.css'
+import 'vuetify/styles'
+
+import { createVuetify, ThemeDefinition } from 'vuetify'
+import * as components from 'vuetify/components'
+import * as directives from 'vuetify/directives'
+
+const theme: ThemeDefinition = {
+  dark: false,
+  colors: {
+    background: '#FFFFFF',
+    surface: '#FFFFFF',
+    primary: '#6200EE',
+    'primary-darken-1': '#3700B3',
+    secondary: '#03DAC6',
+    'secondary-darken-1': '#018786',
+    error: '#B00020',
+    info: '#2196F3',
+    success: '#4CAF50',
+    warning: '#FB8C00',
+  },
+}
+
+export default createVuetify({
+  components,
+  directives,
+  theme: {
+    defaultTheme: 'theme',
+    themes: {
+      theme,
+    },
+  },
+})
